+++ doc/examples/spritesheet/lib/main.dart
@@ -1,11 +1,12 @@
+import 'package:flame/components/animation_component.dart';
+import 'package:flame/components/component.dart';
 import 'package:flame/flame.dart';
 import 'package:flame/game.dart';
 import 'package:flame/spritesheet.dart';
-import 'package:flame/components/animation_component.dart';
-import 'package:flame/components/component.dart';
 import 'package:flutter/material.dart';
 
 void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
   final Size size = await Flame.util.initialDimensions();
   final game = MyGame(size);
   runApp(game.widget);