+++ doc/examples/tiled/lib/main.dart
@@ -1,12 +1,13 @@
-import 'package:flame/components/tiled_component.dart';
-import 'package:flame/game.dart';
-import 'package:flame/flame.dart';
 import 'package:flame/animation.dart';
 import 'package:flame/components/animation_component.dart';
+import 'package:flame/components/tiled_component.dart';
+import 'package:flame/flame.dart';
+import 'package:flame/game.dart';
 import 'package:flutter/widgets.dart' hide Animation;
 import 'package:tiled/tiled.dart' show ObjectGroup, TmxObject;
 
 void main() {
+  WidgetsFlutterBinding.ensureInitialized();
   Flame.images.load('coins.png');
   final TiledGame game = TiledGame();
   runApp(game.widget);
@@ -20,17 +21,21 @@ class TiledGame extends BaseGame {
   }
 
   void _addCoinsInMap(TiledComponent tiledMap) async {
-    final ObjectGroup obj =
+    final ObjectGroup objGroup =
         await tiledMap.getObjectGroupFromLayer("AnimatedCoins");
-    if (obj == null) {
+    if (objGroup == null) {
       return;
     }
-    for (TmxObject obj in obj.tmxObjects) {
+    objGroup.tmxObjects.forEach((TmxObject obj) {
       final comp = AnimationComponent(
-          20.0, 20.0, Animation.sequenced('coins.png', 8, textureWidth: 20));
+        20.0,
+        20.0,
+        Animation.sequenced('coins.png', 8,
+            textureWidth: 20, textureHeight: 20),
+      );
       comp.x = obj.x.toDouble();
       comp.y = obj.y.toDouble();
       add(comp);
-    }
+    });
   }
 }